Retrieve full records in the CSW harvester and unify the record parsing with the GeoNetworkQ harvester.
