+++ Cargo.toml
@@ -17,12 +17,12 @@ futures-util = { version = "0.3", default-features = false }
 once_cell = { version = "1.13", features = ["parking_lot"] }
 hashbrown = { version = "0.12", features = ["serde"] }
 parking_lot = "0.12"
-quick-xml = { version = "0.23", features = ["serialize"] }
 rayon = "1.5"
 regex = "1.6"
 reqwest = { version = "0.11", features = ["json"] }
 scraper = { version = "0.13", default-features = false }
 serde = { version = "1.0", features = ["derive"] }
+serde-xml-rs = "0.5"
 tantivy = { version = "0.18", default-features = false, features = ["mmap"] }
 time = { version = "0.3", features = ["formatting", "macros"] }
 tokio = { version = "1.0", features = ["macros", "rt-multi-thread", "fs", "parking_lot"] }
