+++ templates/csw_get_records.xml
@@ -2,12 +2,12 @@
 
 <csw:GetRecords
     xmlns:csw="http://www.opengis.net/cat/csw/2.0.2"
-    xmlns:gmd="http://www.isotc211.org/2005/gmd"
     service="CSW"
     version="2.0.2"
     resultType="results"
+    outputSchema="csw:Record"
     maxRecords="{{ max_records }}"
     startPosition="{{ start_pos }}"
 >
-    <csw:Query typeNames="gmd:MD_Metadata" />
+    <csw:Query typeNames="csw:Record" />
 </csw:GetRecords>
